// ui/packages/react-devkit/src/domain/types/device.ts
export type DeviceLast = {
  err?: string | null
  type?: string
  snap?: ContainerSnap
  alerts?: Alert[] | null
  [key: string]: unknown
}

export type PosHistoryEntry = {
  container: string
  pos: string
  removedAt: number
}
export type DeviceInfo = {
  container?: string
  pos?: string
  poolConfig?: string
  serialNum?: string
  macAddress?: string | null
  posHistory?: Partial<PosHistoryEntry[]>
  [key: string]: unknown
}

export type Device = {
  id: string
  type: string
  tags?: string[]
  rack?: string
  last?: DeviceLast
  username?: string
  info?: DeviceInfo
  containerId?: string
  address?: string | null
  code?: string
  alerts?: Alert[] | null
  powerMeters?: Device[]
  tempSensors?: Device[]
  transformerTempSensor?: Device
  rootTempSensor?: Device
  err?: string | null
  [key: string]: unknown
}
// ...
export type ContainerSnap = {
  stats?: Partial<ContainerStats>
  config?: Record<string, unknown>
}